refactor into status service
