+++ server/src/middleware/auth_middleware.ts
